Tests passing with new method of building tree.
Oddly, there seems to be a rule which allows adding @reverse links
back to already seen nodes _only_ if the parent of those nodes is
a list item: a bug in the existing implementation? There is a test
which covers it, however, so I have included it.

Lotsa debugging information to remove.